LayoutTests/platform/chromium-win/fast/events/panScroll-panIcon-expected.txt

You should run this test in DRT. 
For manual testing, hold middle button in scrollable and move aroudn mouse pointer for scrolling, then release middle button to stop scrolling. You won't see pan icon after pan scroll.
Scrollable
line 0
line 1
line 2
line 3
line 4
line 5
line 6
line 7
line 8
line 9
